requirements.txt
gradio
huggingface_hub
sentence-transformers
faiss-cpu
transformers
pdfplumber
torch  # Version compatible with sentence-transformers and transformers
numpy
